TestRunner: move the classes of the new implementation into a dedicated folder
